backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1